Create an ability to run secondary METS script
We pull out the execution loop form create_aip_mets and then we
prepare to output the tool information separately. Before we do
that however, we make sure everything still works. So the second
version of the scirpt should output the same as the first, but with
the minor refactory in-place to develop out.